Fixed warnings
- added (const char *) attribute to param_u union to avoid casting away
  const warnings
- added also (const void *) and (volatile void *) for the rest
- there was a typo in computing requested_mtu causing violating sequence
  point evaluation order
- alsa_get_pcm_state_name: added attribute unused
- audio/utils.cpp: different signdness
- audio/utils.cpp: different signdness
- split capture filter: useless dispose_frame
- video_compress: not copying whole string (with strncpy)

+++ src/audio/utils.cpp
@@ -184,7 +184,7 @@ void copy_channel(char *out, const char *in, int bps, int in_len /* bytes */, in
 }
 
 void audio_frame_multiply_channel(struct audio_frame *frame, int new_channel_count) {
-        assert(frame->max_size >= (unsigned int) frame->data_len * new_channel_count / frame->ch_count);
+        assert((unsigned int) frame->max_size >= (unsigned int) frame->data_len * new_channel_count / frame->ch_count);
 
         copy_channel(frame->data, frame->data, frame->bps, frame->data_len, new_channel_count);
 }
